perf: avoid using KEYS command of Redis
BREAKING CHANGE: After this upgrade reset() won't work as expected. Because, to avoid performance penalties on Redis we stopped using `KEYS` command. Instead, all the cache-keys are stored in a redis set, and looked-up from it at the time of resetting cache. So, you need to migrate the existing caches manually, or need to cleanupr/reset existing caches manually. Let us know if you need help on the same.